Menambahkan seeder

+++ database/migrations/2025_10_22_110650_create_pengaduan_table.php
@@ -10,10 +10,10 @@ return new class extends Migration {
         Schema::create('pengaduan', function (Blueprint $table) {
             $table->bigIncrements('pengaduan_id');
             $table->string('nomor_tiket')->unique();
-        
-            $table->unsignedInteger('warga_id'); // sama tipe dengan increments()
+
+            $table->unsignedBigInteger('warga_id'); // sama tipe dengan increments()
             $table->unsignedBigInteger('kategori_id');
-        
+
             $table->string('judul');
             $table->text('deskripsi');
             $table->enum('status', ['pending', 'proses', 'selesai'])->default('pending');
@@ -21,18 +21,18 @@ return new class extends Migration {
             $table->string('rt', 5)->nullable();
             $table->string('rw', 5)->nullable();
             $table->timestamps();
-        
+
             // Foreign keys
             $table->foreign('warga_id')
                 ->references('warga_id')
                 ->on('warga')
                 ->onDelete('cascade');
-        
+
             $table->foreign('kategori_id')
                 ->references('id')
                 ->on('kategori_pengaduan')
                 ->onDelete('cascade');
-        });        
+        });
     }
 
     public function down(): void
